+++ src/Exception/FileException.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Hexogen\KDTree\Exception;
+
+/**
+ * Thrown when a persisted tree file cannot be opened, read, written,
+ * or does not have the expected binary format.
+ */
+class FileException extends \RuntimeException
+{
+}